fixed Merkle root circular dependency
